Dashboard
Hello Everyone,
Found this awesome dashboard template that is under an MIT license. Which means permission to use and edit is free of charge and "Granted." Be sure to read the agreement but I think this will be a great first step. Lets discuss further tomorrow during class.

Cheers,
Nathan